# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AddressRequest": {
        "properties": {
          "address_line_1": {
            "maxLength": 180,
            "minLength": 3,
            "title": "Address Line 1",
            "type": "string"
          },
          "address_line_2": {
            "anyOf": [
              {
                "maxLength": 180,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Line 2"
          },
          "district": {
            "maxLength": 120,
            "minLength": 2,
            "title": "District",
            "type": "string"
          },
          "label": {
            "maxLength": 60,
            "minLength": 2,
            "title": "Label",
            "type": "string"
          },
          "locality": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Locality",
            "type": "string"
          },
          "postal_code": {
            "pattern": "^\\d{6}$",
            "title": "Postal Code",
            "type": "string"
          },
          "state": {
            "maxLength": 120,
            "minLength": 2,
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "label",
          "address_line_1",
          "locality",
          "district",
          "state",
          "postal_code"
        ],
        "title": "AddressRequest",
        "type": "object"
      },
      "AddressResponse": {
        "properties": {
          "address_line_1": {
            "maxLength": 180,
            "minLength": 3,
            "title": "Address Line 1",
            "type": "string"
          },
          "address_line_2": {
            "anyOf": [
              {
                "maxLength": 180,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Line 2"
          },
          "district": {
            "maxLength": 120,
            "minLength": 2,
            "title": "District",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "maxLength": 60,
            "minLength": 2,
            "title": "Label",
            "type": "string"
          },
          "locality": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Locality",
            "type": "string"
          },
          "postal_code": {
            "pattern": "^\\d{6}$",
            "title": "Postal Code",
            "type": "string"
          },
          "state": {
            "maxLength": 120,
            "minLength": 2,
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "label",
          "address_line_1",
          "locality",
          "district",
          "state",
          "postal_code",
          "id"
        ],
        "title": "AddressResponse",
        "type": "object"
      },
...
      "EmailLoginRequest": {
        "properties": {
          "client_type": {
            "default": "customer_web",
            "enum": [
              "mobile",
              "customer_web"
            ],
            "title": "Client Type",
            "type": "string"
          },
          "device_label": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Label"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "EmailLoginRequest",
        "type": "object"
      },
      "EmailRegistrationRequest": {
        "properties": {
          "display_name": {
            "maxLength": 100,
            "minLength": 2,
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 10,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "display_name"
        ],
        "title": "EmailRegistrationRequest",
        "type": "object"
      },
      "EmailVerificationRequest": {
        "properties": {
          "challenge_id": {
            "format": "uuid",
            "title": "Challenge Id",
            "type": "string"
          },
          "client_type": {
            "default": "customer_web",
            "enum": [
              "mobile",
              "customer_web"
            ],
            "title": "Client Type",
            "type": "string"
          },
          "code": {
            "pattern": "^\\d{6}$",
            "title": "Code",
            "type": "string"
          },
          "device_label": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Label"
          }
        },
        "required": [
          "challenge_id",
          "code"
        ],
        "title": "EmailVerificationRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MobileOTPRequest": {
        "properties": {
          "mobile_number": {
            "title": "Mobile Number",
            "type": "string"
          }
        },
        "required": [
          "mobile_number"
        ],
        "title": "MobileOTPRequest",
        "type": "object"
      },
...
      "ProfileUpdateRequest": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "home_locality": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Locality"
          },
          "preferred_language": {
            "anyOf": [
              {
                "enum": [
                  "en",
                  "ta"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Language"
          }
        },
        "title": "ProfileUpdateRequest",
        "type": "object"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          }
        },
        "title": "RefreshRequest",
        "type": "object"
      },
...
      "SessionTokensResponse": {
        "properties": {
          "access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Token"
          },
          "csrf_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Csrf Token"
          },
          "expires_in_seconds": {
            "title": "Expires In Seconds",
            "type": "integer"
          },
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "expires_in_seconds",
          "user"
        ],
        "title": "SessionTokensResponse",
        "type": "object"
      },
      "UserResponse": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "preferred_language": {
            "title": "Preferred Language",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "preferred_language",
          "status"
        ],
        "title": "UserResponse",
        "type": "object"
      },
...
      "app__domains__identity__schemas__ChallengeResponse": {
        "properties": {
          "challenge_id": {
            "format": "uuid",
            "title": "Challenge Id",
            "type": "string"
          },
          "expires_in_seconds": {
            "title": "Expires In Seconds",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "challenge_id",
          "expires_in_seconds",
          "message"
        ],
        "title": "ChallengeResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "4by4 For Hire API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/email/register": {
      "post": {
        "operationId": "register_email_api_v1_auth_email_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailRegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__domains__identity__schemas__ChallengeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Email",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/auth/email/verify": {
      "post": {
        "operationId": "verify_email_api_v1_auth_email_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionTokensResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Email",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_email_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionTokensResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login Email",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "current_user_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Current User",
        "tags": [
          "identity"
        ]
      },
      "patch": {
        "operationId": "update_current_user_api_v1_auth_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Current User",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/auth/me/addresses": {
      "get": {
        "operationId": "list_addresses_api_v1_auth_me_addresses_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AddressResponse"
                  },
                  "title": "Response List Addresses Api V1 Auth Me Addresses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Addresses",
        "tags": [
          "identity"
        ]
      },
      "post": {
        "operationId": "create_address_api_v1_auth_me_addresses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Address",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/auth/me/addresses/{address_id}": {
      "delete": {
        "operationId": "delete_address_api_v1_auth_me_addresses__address_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "address_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Address Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Address",
        "tags": [
          "identity"
        ]
      },
      "put": {
        "operationId": "update_address_api_v1_auth_me_addresses__address_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "address_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Address Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Address",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/auth/mobile/request-otp": {
      "post": {
        "operationId": "request_mobile_otp_api_v1_auth_mobile_request_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MobileOTPRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__domains__identity__schemas__ChallengeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Mobile Otp",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/auth/mobile/verify-otp": {
      "post": {
        "operationId": "verify_mobile_otp_api_v1_auth_mobile_verify_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MobileOTPVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionTokensResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Mobile Otp",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/auth/sessions": {
      "get": {
        "operationId": "list_sessions_api_v1_auth_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionResponse"
                  },
                  "title": "Response List Sessions Api V1 Auth Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Sessions",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/auth/sessions/{session_id}": {
      "delete": {
        "operationId": "revoke_session_api_v1_auth_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Session",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/auth/token/refresh": {
      "post": {
        "operationId": "refresh_session_api_v1_auth_token_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionTokensResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Session",
        "tags": [
          "identity"
        ]
      }
    },
...
  }
}